Issue #17. Fixed warnings (part two)

+++ include/arch/pico_posix.h
@@ -36,14 +36,14 @@ static inline unsigned long PICO_TIME(void)
 {
   struct timeval t;
   gettimeofday(&t, NULL);
-  return (t.tv_sec);
+  return (uint32_t)(t.tv_sec);
 }
 
 static inline unsigned long PICO_TIME_MS(void)
 {
   struct timeval t;
   gettimeofday(&t, NULL);
-  return (t.tv_sec * 1000) + (t.tv_usec / 1000);
+  return (uint32_t)((t.tv_sec * 1000) + (t.tv_usec / 1000));
 }
 
 static inline void PICO_IDLE(void)
